+++ woonuxt_base/components/WishListItem.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 const { removeFromWishlist } = useWishlist();
-const { formatURI } = useHelpers();
 const { product } = defineProps<{ product: Product }>();
 </script>
 
@@ -11,7 +10,7 @@ const { product } = defineProps<{ product: Product }>();
         <path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="32" d="M368 368L144 144M368 144L144 368" />
       </svg>
     </button>
-    <NuxtLink v-if="product.slug" :to="`/product/${formatURI(product.slug)}`">
+    <NuxtLink v-if="product.slug" :to="`/product/${decodeURIComponent(product.slug)}`">
       <img
         v-if="product?.image?.cartSourceUrl"
         class="rounded-lg object-cover h-20 w-20"
@@ -21,7 +20,7 @@ const { product } = defineProps<{ product: Product }>();
         width="100"
         height="120" />
     </NuxtLink>
-    <NuxtLink v-if="product.slug" class="text-lg leading-tight" :to="`/product/${formatURI(product.slug)}`">{{ product.name }}</NuxtLink>
+    <NuxtLink v-if="product.slug" class="text-lg leading-tight" :to="`/product/${decodeURIComponent(product.slug)}`">{{ product.name }}</NuxtLink>
     <ProductPrice :sale-price="product.salePrice" :regular-price="product.regularPrice" class="ml-auto" />
   </li>
 </template>